Checksum... - change return type to uint16 - fix some typo's - fix a bug that stopped us not being able to calculate correctly for odd   values.
Thanks to Bob for pointing out the odd vs even thing on IRC.

+++ net_kit/source/server/include/net_misc.h
@@ -38,7 +38,7 @@ typedef	uint32	ipv4_addr;
  */
 
 void net_server_add_device(ifnet *ifn);
-int in_cksum(struct mbuf *m, int len, int off);
+uint16 in_cksum(struct mbuf *m, int len, int off);
 void local_init(void);
 void insert_local_address(struct sockaddr *sa, ifnet *dev);
 int is_address_local(void *ptr, int len);
